Extract statusbaritem
